requirements.txt
allure-pytest~=2.13.5
Appium-Python-Client~=4.0.0
pytest~=8.1.1
selene~=2.0.0rc9
selenium~=4.19.0